requirements.txt
dataclasses-json==0.6.6
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.8.0
iniconfig==2.0.0
marshmallow==3.21.3
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
python-dateutil==2.9.0.post0
six==1.16.0
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.12.1
